fixed healthCredits type, added decimal library, changed data type for refunds

+++ packages/api/src/profiles-service/resolvers/saveMedicineOrderPaymentmq.ts
@@ -64,7 +64,7 @@ export const saveMedicineOrderPaymentMqTypeDefs = gql`
     CODCity: CODCity
     orderId: String
     paymentMode: PAYMENT_METHODS
-    healthCredits: Int
+    healthCredits: Float
   }
 
   type SaveMedicineOrderPaymentMqResult {
